discordhandlerをハブとして処理ごとに細分化しました
