+++ discordHelloHandler/commands/index.mjs
@@ -1,0 +1,2 @@
+export { handleSlashCommand } from "./slashCommands.js";
+export { handleComponentInteraction } from "./interactionComponents.js";
