+++ discordHelloHandler/commands/subjectMake.mjs
@@ -1,0 +1,42 @@
+import { respondJSON, sendLogMessage } from "../utils.js";
+import AWS from "aws-sdk";
+
+const lambda = new AWS.Lambda({ region: "ap-northeast-1" });
+
+export async function handleSubjectMakeCommand(body, botToken, logChannelId) {
+  const name = body.data.options?.find((opt) => opt.name === "name")?.value;
+  if (!name) {
+    return respondJSON({
+      type: 4,
+      data: { content: "名前が指定されていません。" },
+    });
+  }
+
+  // 例: DynamoDB書き込み用にLambdaを呼ぶ
+  const payload = { table: "sub", track: "new", name };
+
+  try {
+    await lambda
+      .invoke({
+        FunctionName: "arn:aws:lambda:ap-northeast-1:xxx:function:Terakoya_DynamoDB_Write",
+        InvocationType: "Event",
+        Payload: JSON.stringify(payload),
+      })
+      .promise();
+
+    await sendLogMessage(body, botToken, logChannelId);
+
+    return respondJSON({
+      type: 4,
+      data: {
+        content: `subjectテーブルに「${name}」が登録されました。`,
+      },
+    });
+  } catch (error) {
+    console.error("Lambda呼び出しエラー:", error);
+    return respondJSON({
+      type: 4,
+      data: { content: "データ登録中にエラーが発生しました。" },
+    });
+  }
+}
